+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## next
+* Add `Semigroup` instances for `HTML` and `Context` on GHC 8.0+
+
 ## 0.2.1
 * Replace deprecated usages of `apply`, `Translate`, and `translate` with `applyT`, `Transform`, and `transform`, respectively.